add pytest fixtures for RDM records and notification data
